backend/requirements.txt
# FastAPI & server
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pgvector==0.3.2

# Background tasks
celery[redis]==5.4.0
redis==5.0.7

# Pydantic & validation
pydantic==2.7.4
pydantic-settings==2.3.4

# AI / LLM
ollama==0.2.1
openai-whisper>=20231117
# File storage
supabase==2.5.1
pypdf==4.2.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0

# File uploads
python-multipart==0.0.9

# Utilities
httpx==0.27.0
python-dotenv==1.0.1
